+++ bamboost/common/job.py
@@ -52,7 +52,7 @@ class Job:
             script += f"#SBATCH --tmp={tmp}\n"
         script += f"#SBATCH --output={os.path.join(path, uid)}/{uid}.out\n"
 
-        # Add SCRIPT_DIR as environment variable
+        # add SCRIPT_DIR as environment variable
         script += 'SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )\n\n'
 
         # user defined commands
@@ -78,7 +78,7 @@ class Job:
 
         script = f"#!/bin/bash\n\n"
 
-        # Add SCRIPT_DIR as environment variable
+        # add SCRIPT_DIR as environment variable
         script += 'SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )\n\n'
 
         # user defined commands